bmt, file: Move asynchasher to file/hasher
